Time measurement

+++ include/record.hpp
@@ -10,7 +10,7 @@ namespace web_crawler {
         private:
             std::string domain;
             int visited_pages = 0;
-            float total_time = 0.0;
+            int total_time = 0;
             bool currently_visiting = false;
             std::map<std::string, int> urls;
             bool valid(std::string url);
@@ -18,12 +18,12 @@ namespace web_crawler {
         public:
             Record(std::string domain);
             int get_visited_pages();
-            float get_total_time();
+            int get_total_time();
             bool is_currently_visiting();
             bool has_visited(std::string url);
             bool can_start_visiting(std::string url);
             void start_visiting(std::string url);
-            void visit_page(std::string url, float time);
+            void visit_page(std::string url, int time);
             void finish();
     };
 }
